[REFACTOR] View를 VC로 부터 분리 작업
